==> 7_2_subset_imputed_diseased.ipynb <==

# Preprocessing - Integration
Michael Sterr

2024-05-18 


# Setup


# General
import scipy as sci
import numpy as np
import pandas as pd
import logging
import time
import pickle
from itertools import chain
import session_info
import gc # Free memory #gc.collect()
import scipy.stats as stats


# Analysis
import scvi

import scanpy as sc
import anndata as ad

# Load Data

adata_dca = sc.read_h5ad('/mnt/hdd/data/Diseased/Dbtl_detected_velocyto_scran_diseased_sct_imputed.h5ad')

adata = sc.read_h5ad('/mnt/hdd/data/Diseased/adata_markedDoublets_normalized_initialAnno_noimmune_scvi_wodblts.h5ad')

bc_wo_dblts = [name for name in adata.obs_names]

bc_wo_dblts

adata_dca = adata_dca[bc_wo_dblts]

vars_dca = [name for name in adata_dca.var_names]

adata

adata_dca

adata = adata[:,vars_dca]

adata_dca.write('/mnt/hdd/data/Diseased/Dbtl_detected_velocyto_scran_diseased_sct_imputed_subsetted.h5ad')

import h5py

# add dca imputed counts
from anndata._io.specs import read_elem
with h5py.File('/mnt/hdd/data/Diseased/Dbtl_detected_velocyto_scran_diseased_sct_imputed_subsetted.h5ad', 'r') as f:
    # Read specific columns from `obs`
    #sample_column = f['obs/sample'][:]
    #n_counts_column = f['obs/n_counts'][:]
    #https://github.com/scverse/anndata/issues/436:
    #cell_types = read_elem(f["obs/celltype"])
    #umap = read_elem(f["obsm/X_umap"])
    logsct = read_elem(f["layers/sct_logcounts"])
    logdca = read_elem(f['layers/log_dca_counts'])
    spliced = read_elem(f['layers/spliced'])
    unspliced = read_elem(f['layers/unspliced'])
    ambiguous = read_elem(f['layers/ambiguous'])

adata.layers['sct_logcounts'] = logsct
adata.layers['log_dca_counts'] = logdca
adata.layers['spliced'] = spliced
adata.layers['unspliced'] = unspliced
adata.layers['ambiguous'] = ambiguous

del adata_dca
gc.collect()

adata.write('/mnt/hdd/data/Diseased/adata_markedDoublets_normalized_initialAnno_noimmune_scvi_wodblts_imputed_subsetted.h5ad')
